# docs/manager/swagger.yaml
basePath: /
definitions:
  main.errorResponse:
    properties:
      error:
        type: string
    type: object
info:
  contact: {}
  description: Management API for users, voicebots, providers, models, MCP servers,
    memory, and knowledge bases.
  title: Orion X Manager API
  version: "1.0"
paths:
  /api/api-keys:
    get:
      description: Only masked keys are returned, never the plaintext.
      parameters:
      - description: Page number (default 1)
        in: query
        name: page
        type: integer
      - description: Page size (default 20, max 100)
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/main.errorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/main.errorResponse'
      security:
      - BearerAuth: []
      summary: List API keys
      tags:
      - API Keys
    post:
      consumes:
      - application/json
      description: Returns the plaintext key exactly once; it cannot be retrieved
        later.
      parameters:
      - description: name, optional expires_at, and either scopes or preset
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/main.errorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/main.errorResponse'
      security:
      - BearerAuth: []
      summary: Create an API key
      tags:
      - API Keys
  /api/api-keys/{id}:
    delete:
      description: Soft delete; idempotent. Revocation is effective immediately.
      parameters:
      - description: API key ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/main.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/main.errorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/main.errorResponse'
      security:
      - BearerAuth: []
      summary: Revoke an API key
      tags:
      - API Keys
  /api/api-keys/scopes:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/main.errorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/main.errorResponse'
      security:
      - BearerAuth: []
      summary: List API key scopes and presets
      tags:
      - API Keys
  /api/assets:
    get:
      parameters:
      - description: Filter by purpose (image | voice:sample | kb:document)
        in: query
        name: purpose
        type: string
      - description: Filter by file name
        in: query
        name: q
        type: string
      - description: Page number, starting at 1
        in: query
        name: page
        type: integer
      - description: Page size (default 20, max 100)
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/main.errorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/main.errorResponse'
      security:
      - BearerAuth: []
      summary: List resources
      tags:
      - Assets
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: Resource file
        in: formData
        name: file
        required: true
        type: file
      - description: 'Purpose: image | voice:sample | kb:document'
        in: formData
        name: purpose
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/main.errorResponse'
        "413":
          description: Request Entity Too Large
          schema:
            $ref: '#/definitions/main.errorResponse'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/main.errorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/main.errorResponse'
      security:
      - BearerAuth: []
      summary: Upload resource
      tags:
      - Assets
  /api/assets/{id}:
    delete:
      parameters:
      - description: Asset ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/main.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/main.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/main.errorResponse'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/main.errorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/main.errorResponse'
      security:
      - BearerAuth: []
      summary: Delete resource
      tags:
      - Assets
    get:
      parameters:
      - description: Asset ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/main.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/main.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/main.errorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/main.errorResponse'
      security:
      - BearerAuth: []
      summary: Get resource
      tags:
      - Assets
  /api/assets/{id}/url:
    get:
      parameters:
      - description: Asset ID
        in: path
        name: id
        required: true
        type: string
      - description: Set to 1 for a download URL with file name
        in: query
        name: download
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/main.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/main.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/main.errorResponse'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/main.errorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/main.errorResponse'
      security:
      - BearerAuth: []
      summary: Presign resource URL
      tags:
      - Assets
  /api/auth/bind-email:
    post:
      consumes:
      - application/json
      parameters:
      - description: Email address
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/main.errorResponse'
      security:
      - BearerAuth: []
      summary: Bind email
      tags:
      - Authentication
  /api/auth/change-password:
    post:
      consumes:
      - application/json
      parameters:
      - description: Current and new password
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/main.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/main.errorResponse'
      security:
      - BearerAuth: []
      summary: Change password
      tags:
      - Authentication
  /api/auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Username and password
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/main.errorResponse'
      summary: Log in
      tags:
      - Authentication
  /api/auth/profile:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/main.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/main.errorResponse'
      security:
      - BearerAuth: []
      summary: Get current profile
      tags:
      - Authentication
  /api/available-resources:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/main.errorResponse'
      security:
      - BearerAuth: []
      summary: List available resources
      tags:
      - Resources
  /api/channels:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: object
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/main.errorResponse'
      security:
      - BearerAuth: []
      summary: List channel platforms
      tags:
      - Channels
  /api/data/knowledge/bots/{bot_id}/knowledge_bases:
    get:
      parameters:
      - description: Voicebot ID
        in: path
        name: bot_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/main.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/main.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/main.errorResponse'
      security:
      - BearerAuth: []
      summary: List bot knowledge bases
      tags:
      - Knowledge
    post:
      consumes:
      - application/json
      parameters:
      - description: Voicebot ID
        in: path
        name: bot_id
        required: true
        type: string
      - description: Knowledge base fields
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/main.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/main.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/main.errorResponse'
      security:
      - BearerAuth: []
      summary: Create bot knowledge base
      tags:
      - Knowledge
  /api/data/knowledge/bots/{bot_id}/knowledge_bases/{kb_id}/bind:
    delete:
      parameters:
      - description: Voicebot ID
        in: path
        name: bot_id
        required: true
        type: string
      - description: Knowledge base ID
        in: path
        name: kb_id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/main.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/main.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/main.errorResponse'
      security:
      - BearerAuth: []
      summary: Unbind knowledge base
      tags:
      - Knowledge
  /api/data/knowledge/bots/{bot_id}/knowledge_bases/bind:
    post:
      consumes:
      - application/json
      parameters:
      - description: Voicebot ID
        in: path
        name: bot_id
        required: true
        type: string
      - description: Knowledge base ID
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/main.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/main.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/main.errorResponse'
      security:
      - BearerAuth: []
      summary: Bind knowledge base
      tags:
      - Knowledge
  /api/data/knowledge/bots/{bot_id}/knowledge_bases/bound:
    get:
      parameters:
      - description: Voicebot ID
        in: path
        name: bot_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/main.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/main.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/main.errorResponse'
      security:
      - BearerAuth: []
      summary: List bound knowledge bases
      tags:
      - Knowledge
  /api/data/knowledge/documents/{doc_id}:
    delete:
      parameters:
      - description: Document ID
        in: path
        name: doc_id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/main.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/main.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/main.errorResponse'
      security:
      - BearerAuth: []
      summary: Delete knowledge document
      tags:
      - Knowledge
  /api/data/knowledge/documents/{doc_id}/retry:
    post:
      parameters:
      - description: Document ID
        in: path
        name: doc_id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/main.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/main.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/main.errorResponse'
      security:
      - BearerAuth: []
      summary: Retry document processing
      tags:
      - Knowledge
  /api/data/knowledge/documents/{doc_id}/status:
    get:
      parameters:
      - description: Document ID
        in: path
        name: doc_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/main.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/main.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/main.errorResponse'
      security:
      - BearerAuth: []
      summary: Get document status
      tags:
      - Knowledge
  /api/data/knowledge/knowledge_bases:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/main.errorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/main.errorResponse'
      security:
      - BearerAuth: []
      summary: List knowledge bases
      tags:
      - Knowledge
...
  /api/data/memory/{id}:
    delete:
      parameters:
      - description: Memory entry ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/main.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/main.errorResponse'
      security:
      - BearerAuth: []
      summary: Delete memory entry
      tags:
      - Memory
...
securityDefinitions:
  BearerAuth:
    description: Enter a JWT as `Bearer <token>`.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"